Math Product website
It was a Math game project in which a multiplication and respective four answer is generating in 60 sec countdown, where 1 correct answer and remainning 3 answer are wrong from which you have to choose 1 option. If it is correct then it shows a dialog box of coorect otherwise it shows try again and your score will reflect in score box. At the end of countdown it shows a dailog box of Game over with your final score. Actually in 60 sec how much you scored is showing your efficiency in math.